[DM Tool] - Created class DotTrace which reads elf and trace files, partitions the complete application into InstructionBlocks and writes a DOT representation of the trace; - Added method autoSet to class Settings;- CVS uses 'tab' as separator instead of ',', because it is the default setting in Excel; - In package Stats.TraceCoverage, changed some Integer types to Long; - In DmPartitionerDispenser, removed getDaprofSimple and changed getDaprof to getWarp; - Cleaned classes Transform, WriteBlocks, Shell and TraceCoverage;- Removed option "partition.daprofuseoriginalidmethod"; - Changed "partition.daprofusebranchlimit" to "partition.warpusebranchlimit"; - Deleted package Shell.System.Transform;

+++ src/dm-tool/org/ancora/DMTool/Stats/TraceCoverage/TcBlocksizeData.java
@@ -29,14 +29,14 @@ import org.ancora.InstructionBlock.InstructionBlock;
 public class TcBlocksizeData {
 
    public TcBlocksizeData(String filename) {
-      instPerBlockTotalSize = new HashMap<Integer, Long>();
+      instPerBlockTotalSize = new HashMap<Long, Long>();
       totalInstructions = 0;
       this.filename = filename;
    }
 
 
    public void addBlock(InstructionBlock block) {
-      int totalBlockSize = block.getTotalInstructions();
+      long totalBlockSize = block.getTotalInstructions();
       Long blockInst = instPerBlockTotalSize.get(totalBlockSize);
       if(blockInst == null) {
          blockInst = 0l;
@@ -60,7 +60,7 @@ public class TcBlocksizeData {
    }
 
 
-   public Map<Integer, Long> getInstPerBlockTotalSize() {
+   public Map<Long, Long> getInstPerBlockTotalSize() {
       return instPerBlockTotalSize;
    }
 
@@ -69,7 +69,7 @@ public class TcBlocksizeData {
    /**
     * INSTANCE VARIABLES
     */
-   private Map<Integer, Long> instPerBlockTotalSize;
+   private Map<Long, Long> instPerBlockTotalSize;
    private long totalInstructions;
    private String filename;
 }
